requirements.txt
Django~=4.0.1
django-colorfield
django-markdownx==4.0.0b1